+++ tests/git/Command/GitClone/CloneCmdTest.php
@@ -2,25 +2,25 @@
 
 declare(strict_types=1);
 
-namespace SebastianFeldmann\Git\Command\GitClone;
+namespace SebastianFeldmann\Git\Command\CloneCmd;
 
 use PHPUnit\Framework\TestCase;
-use SebastianFeldmann\Git\Command\GitClone\GitClone;
+use SebastianFeldmann\Git\Command\CloneCmd\CloneCmd;
 
-final class GitCloneTest extends TestCase
+final class CloneCmdTest extends TestCase
 {
     /**
      * Tests GitClone::getCommand
      */
     public function testGitCloneCommand()
     {
-        $clone = new GitClone('https://github.com/test/repo.git');
+        $clone = new CloneCmd('https://github.com/test/repo.git');
         $this->assertEquals('git clone \'https://github.com/test/repo.git\' \'repo\'', $clone->getCommand());
     }
 
     public function testGitCloneCommandDir()
     {
-        $clone = new GitClone('https://github.com/test/repo.git');
+        $clone = new CloneCmd('https://github.com/test/repo.git');
         $clone = $clone->dir('fubar');
         $this->assertEquals('git clone \'https://github.com/test/repo.git\' \'fubar\'', $clone->getCommand());
     }
